# distsql: Refactor aggregator into a state machine