This lays the groundwork for extending the aggregator to stream results
if some of the group columns have an ordering.

Release note: None

----

This is the first commit from #24113, extracted into its own PR.