@asubiotto 2018-04-06
Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---

*[pkg/sql/distsqlrun/aggregator.go, line 360 at r1](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573#-L9Ru2421gJn4ML6m8An:-L9Ru2421gJn4ML6m8Ao:bmcjpu2) ([raw file](https://github.com/cockroachdb/cockroach/blob/e92426e466915fd4df7fd37ad71dbf9d473c7ed2/pkg/sql/distsqlrun/aggregator.go#L360)):*
> ```Go
> 		var row sqlbase.EncDatumRow
> 		var meta *ProducerMetadata
> 		switch ag.state {
> ```

It seems that these three functions have the same signature. Have you considered having these implement an interface? The aggregator could then have a field of this interface that it delegates to instead of the switch statement. This field would be overwritten to point to the next stage once a stage is done.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573)*
<!-- Sent from Reviewable.io -->