@asubiotto 2018-04-06
Review status: 0 of 1 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[pkg/sql/distsqlrun/aggregator.go, line 360 at r1](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573#-L9Ru2421gJn4ML6m8An:-L9S3lVLAHvBQBglAFys:byos89k) ([raw file](https://github.com/cockroachdb/cockroach/blob/e92426e466915fd4df7fd37ad71dbf9d473c7ed2/pkg/sql/distsqlrun/aggregator.go#L360)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

It seems that these three functions have the same signature. Have you considered having these implement an interface? The aggregator could then have a field of this interface that it delegates to instead of the switch statement. This field would be overwritten to point to the next stage once a stage is done.
</blockquote></details>

Sorry, I've been thinking about interfaces too much. What I mean is have a member of the aggregator with the function signature and assign that with these depending on the stage.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573)*
<!-- Sent from Reviewable.io -->