@petermattis 2018-04-07
Review status: 0 of 1 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[pkg/sql/distsqlrun/aggregator.go, line 360 at r1](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573#-L9Ru2421gJn4ML6m8An:-L9VDqnUDq9Gl7ILaqli:b-2c0qln) ([raw file](https://github.com/cockroachdb/cockroach/blob/e92426e466915fd4df7fd37ad71dbf9d473c7ed2/pkg/sql/distsqlrun/aggregator.go#L360)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

Sorry, I've been thinking about interfaces too much. What I mean is have a member of the aggregator with the function signature and assign that with these depending on the stage.
</blockquote></details>

The switch statement might be faster that an indirect call through a function pointer (the CPU branch prediction should predict the case taken very reliably). Not sure if it matters here or not, but something to pay attention to.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573)*
<!-- Sent from Reviewable.io -->