@abhimadan 2018-04-09
Review status: 0 of 1 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[pkg/sql/distsqlrun/aggregator.go, line 360 at r1](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573#-L9Ru2421gJn4ML6m8An:-L9fIx-J5-UgwlOtecXI:b9qcju) ([raw file](https://github.com/cockroachdb/cockroach/blob/e92426e466915fd4df7fd37ad71dbf9d473c7ed2/pkg/sql/distsqlrun/aggregator.go#L360)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

The switch statement might be faster that an indirect call through a function pointer (the CPU branch prediction should predict the case taken very reliably). Not sure if it matters here or not, but something to pay attention to.
</blockquote></details>

Here's the `benchstat` output from comparing the switch statement to the function pointer version:
```name                    old time/op    new time/op    delta
Aggregation/IDENT-8       54.9µs ± 3%    52.6µs ± 2%   -4.26%  (p=0.000 n=10+9)
Aggregation/AVG-8         80.2µs ± 2%    78.6µs ± 1%   -1.93%  (p=0.000 n=9+9)
Aggregation/COUNT-8       54.2µs ± 2%    54.4µs ± 4%     ~     (p=0.481 n=10+10)
Aggregation/MAX-8         61.3µs ± 6%    61.2µs ± 1%     ~     (p=0.353 n=10+10)
Aggregation/MIN-8         59.6µs ± 1%    61.4µs ± 3%   +2.95%  (p=0.001 n=10+10)
Aggregation/STDDEV-8      1.76ms ± 5%    1.74ms ± 2%     ~     (p=0.315 n=10+8)
Aggregation/SUM-8         68.6µs ± 1%    71.4µs ± 4%   +4.12%  (p=0.000 n=9+10)
Aggregation/SUM_INT-8     67.1µs ± 1%    70.2µs ± 4%   +4.64%  (p=0.000 n=10+10)
Aggregation/VARIANCE-8    1.64ms ± 2%    1.68ms ± 1%   +2.00%  (p=0.000 n=10+8)
Aggregation/XOR_AGG-8     53.4µs ± 1%    54.1µs ± 1%   +1.24%  (p=0.000 n=9+9)

name                    old speed      new speed      delta
Aggregation/IDENT-8      146MB/s ± 3%   152MB/s ± 2%   +4.39%  (p=0.000 n=10+9)
Aggregation/AVG-8        100MB/s ± 2%   102MB/s ± 1%   +1.97%  (p=0.000 n=9+9)
Aggregation/COUNT-8      148MB/s ± 2%   147MB/s ± 4%     ~     (p=0.481 n=10+10)
Aggregation/MAX-8        131MB/s ± 6%   131MB/s ± 1%     ~     (p=0.353 n=10+10)
Aggregation/MIN-8        134MB/s ± 1%   130MB/s ± 3%   -2.84%  (p=0.001 n=10+10)
Aggregation/STDDEV-8    4.56MB/s ± 5%  4.60MB/s ± 2%     ~     (p=0.283 n=10+8)
Aggregation/SUM-8        117MB/s ± 1%   112MB/s ± 4%   -3.93%  (p=0.000 n=9+10)
Aggregation/SUM_INT-8    119MB/s ± 1%   114MB/s ± 4%   -4.39%  (p=0.000 n=10+10)
Aggregation/VARIANCE-8  4.87MB/s ± 2%  4.76MB/s ± 2%   -2.23%  (p=0.000 n=10+9)
Aggregation/XOR_AGG-8    150MB/s ± 1%   148MB/s ± 1%   -1.23%  (p=0.000 n=9+9)

name                    old alloc/op   new alloc/op   delta
Aggregation/IDENT-8       3.10kB ± 0%    3.14kB ± 0%   +1.55%  (p=0.000 n=10+10)
Aggregation/AVG-8         3.87kB ± 0%    3.92kB ± 0%   +1.24%  (p=0.000 n=10+10)
Aggregation/COUNT-8       3.29kB ± 0%    3.34kB ± 0%   +1.46%  (p=0.000 n=10+10)
Aggregation/MAX-8         3.31kB ± 0%    3.36kB ± 0%   +1.45%  (p=0.000 n=10+10)
Aggregation/MIN-8         3.31kB ± 0%    3.36kB ± 0%   +1.45%  (p=0.000 n=10+10)
Aggregation/STDDEV-8       407kB ± 0%     407kB ± 0%   +0.01%  (p=0.000 n=9+10)
Aggregation/SUM-8         3.47kB ± 0%    3.52kB ± 0%   +1.38%  (p=0.000 n=10+10)
Aggregation/SUM_INT-8     3.30kB ± 0%    3.34kB ± 0%   +1.46%  (p=0.000 n=10+10)
Aggregation/VARIANCE-8     404kB ± 0%     404kB ± 0%   +0.01%  (p=0.000 n=10+9)
Aggregation/XOR_AGG-8     3.31kB ± 0%    3.36kB ± 0%   +1.45%  (p=0.000 n=10+10)

name                    old allocs/op  new allocs/op  delta
Aggregation/IDENT-8         16.0 ± 0%      19.0 ± 0%  +18.75%  (p=0.000 n=10+10)
Aggregation/AVG-8           33.0 ± 0%      36.0 ± 0%   +9.09%  (p=0.000 n=10+10)
Aggregation/COUNT-8         23.0 ± 0%      26.0 ± 0%  +13.04%  (p=0.000 n=10+10)
Aggregation/MAX-8           23.0 ± 0%      26.0 ± 0%  +13.04%  (p=0.000 n=10+10)
Aggregation/MIN-8           23.0 ± 0%      26.0 ± 0%  +13.04%  (p=0.000 n=10+10)
Aggregation/STDDEV-8       9.11k ± 0%     9.11k ± 0%   +0.03%  (p=0.000 n=10+10)
Aggregation/SUM-8           24.0 ± 0%      27.0 ± 0%  +12.50%  (p=0.000 n=10+10)
Aggregation/SUM_INT-8       23.0 ± 0%      26.0 ± 0%  +13.04%  (p=0.000 n=10+10)
Aggregation/VARIANCE-8     9.04k ± 0%     9.04k ± 0%   +0.03%  (p=0.000 n=10+10)
Aggregation/XOR_AGG-8       24.0 ± 0%      27.0 ± 0%  +12.50%  (p=0.000 n=10+10)
```
So there doesn't seem to be a significant speed difference, though there is a slight (expected) bump in memory allocations.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24573)*
<!-- Sent from Reviewable.io -->